requirements.txt
fastapi
uvicorn
markdown
html2text
jinja2
deta
python-dotenv
aiofiles
python-multipart